requirements.txt
sqlalchemy==1.3.22
psycopg2==2.8.4
PyGithub==1.54.1
